+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   need to prefix each addons with `-m `. Now it could be done like:
   `odoo-helper test my_addon1 my_addon2`
 - Added `--skip` option of `odoo-helper test` command
+- Added `odoo-helper db copy` command
 
 ### Fixed